services/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-dotenv
qrcode[pil]
Pillow
loguru
passlib

